requirements_metal.txt
# macOS Apple Silicon (M1/M2/M3/M4) 专用依赖
# 使用 MPS (Metal Performance Shaders) 进行 GPU 加速
#
# 安装方式:
#   pip install -r requirements_metal.txt
#   pip install git+https://github.com/lucasb-eyer/pydensecrf.git
#
# 注意: 不要从 pytorch.org 的 CUDA 源安装 PyTorch

# PyTorch (使用默认源，自动安装 MPS 版本)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 核心依赖
aiofiles==25.1.0
aiohttp==3.12.15
aioshutil==1.5
albumentations==2.0.8
arabic-reshaper==3.0.0
backports.statistics==0.1.0
colorama==0.4.6
ctranslate2==3.24.0
customtkinter==5.2.2
einops==0.8.1
fastapi==0.116.1
fire==0.7.1
freetype-py==2.5.1
fugashi==1.3.2
python-multipart==0.0.20

httpcore==1.0.9
httpx>=0.28.1,<1.0.0
huggingface-hub==0.34.4
ijson==3.4.0
jaconv==0.3.4
json5==0.9.28
kornia==0.8.1
langcodes==3.5.0
loguru==0.7.3
matplotlib==3.10.5
networkx==3.5
numpy>=2.0,<2.3
omegaconf==2.3.0

# ONNX Runtime (CPU 版本，macOS 不支持 GPU 版)
onnxruntime==1.20.1

openai==2.14.0
curl_cffi>=0.7.0
opencv-contrib-python>=4.10.0
packaging<25.0
pandas==2.3.1
Pillow==11.3.0
pillow-heif==0.21.0
psutil==5.9.5
py3langid==0.2.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pyperclip==1.9.0
uharfbuzz>=0.40.0
pythainlp[nlpo3]==5.3.1

# pydensecrf - 预编译 wheel 文件 (Python 3.12 macOS universal2)
pydensecrf @ https://github.com/hgmzhn/pydensecrf/releases/download/v1.0.0/pydensecrf-1.0-cp312-cp312-macosx_10_13_universal2.whl
# 如果下载失败,可从源码安装(需要 Xcode 命令行工具):
# pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git

PyHyphen==4.0.4
PyQt6==6.9.1
pytest==7.2.2
python-dotenv==1.1.1
rich==14.1.0
safetensors==0.6.2
spandrel==0.4.1
scipy==1.16.1
# YOLO OBB / MangaLens 气泡检测运行时依赖 ultralytics
ultralytics==8.4.14
# sentencepiece 已移除，改用 tokenizers 库 (use_fast=True)
setuptools==80.9.0
Shapely==2.1.1
scikit-image==0.25.2
starlette==0.47.2
tiktoken==0.11.0
timm==1.0.19
transformers==4.55.0
accelerate>=1.12.0
tqdm==4.67.1
unidic_lite==1.0.8
uvicorn==0.35.0
websockets==14.2

# xformers 不支持 macOS，已移除
# xformers - CUDA only, not needed for MPS
anyio>=4.8.0
google-genai==1.57.0
torchsummary
open_clip_torch==3.1.0
python-bidi==0.6.6
tufup==0.9.0
google-cloud==0.34.0

# PDF/EPUB/CBZ 支持
PyMuPDF==1.25.5
rarfile==4.2
py7zr==0.22.0

# Server 模块
bcrypt==4.2.1
cryptography==44.0.0
APScheduler==3.10.4
pytest-anyio==0.0.0
